requirements.txt
django~=3.2
djangorestframework~=3.13.1
django-isbn-field~=0.5.2
gunicorn
django-heroku